side project 2 (WIP)
On-going side project - the end goal is to assemble a tiny 5-inch display (powered by a raspberry pi 3), that can chart the trajectory that the ISS will follow in its orbit around the earth. Code works just fine, but the projects currently on the backburner, due to courses taking more time and the proper hardware being hard to find.